+++ pysys-examples/internal/testcases/PySys_internal_125/run.py
@@ -80,6 +80,17 @@ class PySysTest(BaseTest):
 		self.assertThat('float(startupTime) < 60.0', 
 			startupTime__eval="self.getExprFromFile('myprocess-2.log', 'Server started in ([0-9.]+) seconds')")
 
+		self.assertThat('serverStartInfo == expected', expected={
+			'startupTime':'20.3',
+			'user':None,
+			},
+			serverStartInfo__eval="self.getExprFromFile('myprocess-2.log', 'Server started in (?P<startupTime>[0-9.]+) seconds(?P<user> as user .*)?')")
+		self.assertThat('serverStartInfo == expected', expected=[{
+			'startupTime':'20.3',
+			'user':None,
+			}],
+			serverStartInfo__eval="self.getExprFromFile('myprocess-2.log', 'Server started in (?P<startupTime>[0-9.]+) seconds(?P<user> as user .*)?', returnAll=True)")
+			
 		user = 'myuser'
 		self.assertThat('actualUser == expected', expected='myuser', actualUser=user)